Automatically backed up by Learn

+++ lib/cash_register.rb
@@ -17,7 +17,7 @@ class CashRegister
     if @discount > 0 
       discount1 = (1 -(@discount.to_f/100))
       @total = @total*discount1
-      @total = @total.to_f
+      @total = @total.to_i
       return "After the discount, the total comes to #{@total}."
     else 
       return "There is no discount to apply."
